+++ Assets/Adjust/Native/iOS/AdjustUnity.mm
@@ -434,10 +434,57 @@ extern "C"
         }];
     }
 
+    void _AdjustGetAttributionWithTimeout(int timeoutInMilliseconds, AdjustDelegateAttributionGetter callback) {
+        [Adjust attributionWithTimeout:timeoutInMilliseconds completionHandler:^(ADJAttribution * _Nullable attribution) {
+            if (attribution != nil) {
+                NSMutableDictionary *dictionary = [NSMutableDictionary dictionary];
+                addValueOrEmpty(dictionary, @"trackerToken", attribution.trackerToken);
+                addValueOrEmpty(dictionary, @"trackerName", attribution.trackerName);
+                addValueOrEmpty(dictionary, @"network", attribution.network);
+                addValueOrEmpty(dictionary, @"campaign", attribution.campaign);
+                addValueOrEmpty(dictionary, @"creative", attribution.creative);
+                addValueOrEmpty(dictionary, @"adgroup", attribution.adgroup);
+                addValueOrEmpty(dictionary, @"clickLabel", attribution.clickLabel);
+                addValueOrEmpty(dictionary, @"costType", attribution.costType);
+                addValueOrEmpty(dictionary, @"costAmount", attribution.costAmount);
+                addValueOrEmpty(dictionary, @"costCurrency", attribution.costCurrency);
+                if (attribution.jsonResponse != nil) {
+                    NSData *jsonData = [NSJSONSerialization dataWithJSONObject:attribution.jsonResponse 
+                                                                       options:0
+                                                                         error:nil];
+                    NSString *strJsonResponse = [[NSString alloc] initWithData:jsonData encoding:NSUTF8StringEncoding];
+                    addValueOrEmpty(dictionary, @"jsonResponse", strJsonResponse);
+                }
+
+                NSData *dataAttribution = [NSJSONSerialization dataWithJSONObject:dictionary
+                                                                          options:0
+                                                                            error:nil];
+                NSString *stringAttribution = [[NSString alloc] initWithBytes:[dataAttribution bytes]
+                                                                       length:[dataAttribution length]
+                                                                     encoding:NSUTF8StringEncoding];
+                const char* attributionCString = [stringAttribution UTF8String];
+                callback(attributionCString);
+            } else {
+                // pass NULL when attribution is nil - C# callback will handle it
+                callback(NULL);
+            }
+        }];
+    }
+
     void _AdjustGetAdid(AdjustDelegateAdidGetter callback) {
         [Adjust adidWithCompletionHandler:^(NSString * _Nullable adid) {
             // TODO: nil checks
             callback([adid UTF8String]);
+        }];
+    }
+
+    void _AdjustGetAdidWithTimeout(int timeoutInMilliseconds, AdjustDelegateAdidGetter callback) {
+        [Adjust adidWithTimeout:timeoutInMilliseconds completionHandler:^(NSString * _Nullable adid) {
+            if (adid != nil) {
+                callback([adid UTF8String]);
+            } else {
+                callback(NULL);
+            }
         }];
     }