+++ Assets/Adjust/Scripts/Adjust.cs
@@ -569,6 +569,22 @@ namespace AdjustSdk
 #endif
         }
 
+        public static void GetAdidWithTimeout(int timeoutInMilliseconds, Action<string> callback)
+        {
+            if (IsEditor())
+            {
+                return;
+            }
+
+#if UNITY_IOS
+            AdjustiOS.GetAdidWithTimeout(timeoutInMilliseconds, callback);
+#elif UNITY_ANDROID
+            AdjustAndroid.GetAdidWithTimeout(timeoutInMilliseconds, callback);
+#else
+            Debug.Log(errorMsgPlatform);
+#endif
+        }
+
         public static void GetAttribution(Action<AdjustAttribution> callback)
         {
             if (IsEditor())
@@ -580,6 +596,22 @@ namespace AdjustSdk
             AdjustiOS.GetAttribution(callback);
 #elif UNITY_ANDROID
             AdjustAndroid.GetAttribution(callback);
+#else
+            Debug.Log(errorMsgPlatform);
+#endif
+        }
+
+        public static void GetAttributionWithTimeout(int timeoutInMilliseconds, Action<AdjustAttribution> callback)
+        {
+            if (IsEditor())
+            {
+                return;
+            }
+
+#if UNITY_IOS
+            AdjustiOS.GetAttributionWithTimeout(timeoutInMilliseconds, callback);
+#elif UNITY_ANDROID
+            AdjustAndroid.GetAttributionWithTimeout(timeoutInMilliseconds, callback);
 #else
             Debug.Log(errorMsgPlatform);
 #endif